src/pyproject.toml
[tool.poetry]
name = "brandplus-api"
version = "0.1.0"
description = ""
authors = ["afeef <janjua.afeef@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pika = "^1.3.2"
pydantic = "^2.0.3"
pydantic-settings = "^2.0.2"
pyjwt = "^2.8.0"
pymongo = {extras = ["srv"], version = "^4.4.1"}
pytest = "^7.4.0"
rollbar = "^0.16.3"
sentry-sdk = {extras = ["fastapi"], version = "^1.28.1"}
uvicorn = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
